v0.2.1: fix dynamic searchParams needed to be awaited
